의도를 드러내기 위해 Phone, AbstractPhone 클래스를 개명하라
