+++ src/main/java/com/johngrib/objects/_10_call/Phone.java
@@ -3,25 +3,23 @@ package com.johngrib.objects._10_call;
 import com.johngrib.objects._02_movie.Money;
 import lombok.Getter;
 
-import java.time.Duration;
+import java.util.ArrayList;
+import java.util.List;
 
-/**
- * 일반 요금제.
- */
-public class Phone extends AbstractPhone {
-  /** 단위 요금. */
-  @Getter
-  private Money amount;
-  @Getter
-  private Duration seconds;
+public abstract class Phone {
 
-  public Phone(Money amount, Duration seconds) {
-    this.amount = amount;
-    this.seconds = seconds;
+  /** 전체 통화 목록 */
+  @Getter
+  private List<Call> calls = new ArrayList<>();
+
+  public Money calculateFee() {
+    Money result = Money.ZERO;
+
+    for (Call call : calls) {
+      result = result.plus(calculateCallFee(call));
+    }
+    return result;
   }
 
-  @Override
-  protected Money calculateCallFee(Call call) {
-    return amount.times(call.getDuration().getSeconds() / seconds.getSeconds());
-  }
+  abstract protected Money calculateCallFee(Call call);
 }
